Refactor the converter into different files
