+++ src/CSnakes.Runtime/PyObjectTypeConverter.cs
@@ -110,115 +110,9 @@ internal partial class PyObjectTypeConverter : TypeConverter
         throw new InvalidCastException($"Attempting to cast {destinationType} from {pyObject.GetPythonType()}");
     }
 
-    private object? ConvertToDictionary(PyObject pyObject, Type destinationType, ITypeDescriptorContext? context, CultureInfo? culture, bool useMappingProtocol = false)
-    {
-        using PyObject items = useMappingProtocol ? new(CPythonAPI.PyMapping_Items(pyObject.GetHandle())) : new(CPythonAPI.PyDict_Items(pyObject.GetHandle()));
-        Type item1Type = destinationType.GetGenericArguments()[0];
-        Type item2Type = destinationType.GetGenericArguments()[1];
-        Type dictType = typeof(Dictionary<,>).MakeGenericType(item1Type, item2Type);
-        IDictionary dict = (IDictionary)Activator.CreateInstance(dictType)!;
-        nint itemsLength = CPythonAPI.PyList_Size(items.GetHandle());
-
-        for (nint i = 0; i < itemsLength; i++)
-        {
-            using PyObject item = new(CPythonAPI.PyList_GetItem(items.GetHandle(), i));
-
-            using PyObject item1 = new(CPythonAPI.PyTuple_GetItem(item.GetHandle(), 0));
-            using PyObject item2 = new(CPythonAPI.PyTuple_GetItem(item.GetHandle(), 1));
-
-            object? convertedItem1 = AsManagedObject(item1Type, item1, context, culture);
-            object? convertedItem2 = AsManagedObject(item2Type, item2, context, culture);
-
-            dict.Add(convertedItem1!, convertedItem2);
-        }
-
-        Type returnType = typeof(ReadOnlyDictionary<,>).MakeGenericType(item1Type, item2Type);
-        return Activator.CreateInstance(returnType, dict);
-    }
-
-    private object? ConvertToTuple(ITypeDescriptorContext? context, CultureInfo? culture, PyObject pyObj, Type destinationType)
-    {
-        var tuplePtr = pyObj.GetHandle();
-
-        // We have to convert the Python values to CLR values, as if we just tried As<object>() it would
-        // not parse the Python type to a CLR type, only to a new Python type.
-        Type[] types = destinationType.GetGenericArguments();
-        object?[] clrValues;
-
-        var tupleValues = new List<PyObject>();
-        for (nint i = 0; i < CPythonAPI.PyTuple_Size(tuplePtr); i++)
-        {
-            PyObject value = new(CPythonAPI.PyTuple_GetItem(tuplePtr, i));
-            tupleValues.Add(value);
-        }
-
-        // If we have more than 8 python values, we are going to have a nested tuple, which we have to unpack.
-        if (tupleValues.Count > 8)
-        {
-            // We are hitting nested tuples here, which will be treated in a different way.
-            object?[] firstSeven = tupleValues.Take(7).Select((p, i) => AsManagedObject(types[i], p, context, culture)).ToArray();
-
-            // Get the rest of the values and convert them to a nested tuple.
-            IEnumerable<PyObject> rest = tupleValues.Skip(7);
-
-            // Back to a Python tuple.
-            using PyObject pyTuple = PyTuple.CreateTuple(rest);
-
-            // Use the decoder pipeline to decode the nested tuple (and its values).
-            // We do this because that means if we have nested nested tuples, they'll be decoded as well.
-            object? nestedTuple = AsManagedObject(types[7], pyTuple, context, culture);
-
-            // Append our nested tuple to the first seven values.
-            clrValues = [.. firstSeven, nestedTuple];
-        }
-        else
-        {
-            clrValues = tupleValues.Select((p, i) => AsManagedObject(types[i], p, context, culture)).ToArray();
-        }
-
-        // Dispose of all the Python values that we captured from the Tuple.
-        foreach (var value in tupleValues)
-        {
-            value.Dispose();
-        }
-
-        ConstructorInfo ctor = destinationType.GetConstructors().First(c => c.GetParameters().Length == clrValues.Length);
-        return (ITuple)ctor.Invoke(clrValues);
-    }
-
     private object? AsManagedObject(Type type, PyObject p, ITypeDescriptorContext? context, CultureInfo? culture)
     {
         return ConvertTo(context, culture, p, type);
-    }
-
-    private object? ConvertToList(PyObject pyObject, Type destinationType, ITypeDescriptorContext? context, CultureInfo? culture)
-    {
-        Type genericArgument = destinationType.GetGenericArguments()[0];
-        Type listType = typeof(List<>).MakeGenericType(genericArgument);
-
-        IList list = (IList)Activator.CreateInstance(listType)!;
-        for (var i = 0; i < CPythonAPI.PyList_Size(pyObject.GetHandle()); i++)
-        {
-            using PyObject item = new(CPythonAPI.PyList_GetItem(pyObject.GetHandle(), i));
-            list.Add(AsManagedObject(genericArgument, item, context, culture));
-        }
-
-        return list;
-    }
-
-    private object? ConvertToListFromSequence(PyObject pyObject, Type destinationType, ITypeDescriptorContext? context, CultureInfo? culture)
-    {
-        Type genericArgument = destinationType.GetGenericArguments()[0];
-        Type listType = typeof(List<>).MakeGenericType(genericArgument);
-
-        IList list = (IList)Activator.CreateInstance(listType)!;
-        for (var i = 0; i < CPythonAPI.PySequence_Size(pyObject.GetHandle()); i++)
-        {
-            using PyObject item = new(CPythonAPI.PySequence_GetItem(pyObject.GetHandle(), i));
-            list.Add(AsManagedObject(genericArgument, item, context, culture));
-        }
-
-        return list;
     }
 
     public override object? ConvertFrom(ITypeDescriptorContext? context, CultureInfo? culture, object value) =>
@@ -236,56 +130,6 @@ internal partial class PyObjectTypeConverter : TypeConverter
             null => new PyObject(CPythonAPI.GetNone()),
             _ => base.ConvertFrom(context, culture, value)
         };
-
-    private PyObject ConvertFromDictionary(ITypeDescriptorContext? context, CultureInfo? culture, IDictionary dictionary)
-    {
-        PyObject pyDict = new(CPythonAPI.PyDict_New());
-
-        foreach (DictionaryEntry kvp in dictionary)
-        {
-            int result = CPythonAPI.PyDict_SetItem(pyDict.GetHandle(), ToPython(kvp.Key, context, culture).GetHandle(), ToPython(kvp.Value, context, culture).GetHandle());
-            if (result == -1)
-            {
-                throw new Exception("Failed to set item in dictionary");
-            }
-        }
-
-        return pyDict;
-    }
-
-    private PyObject ConvertFromList(ITypeDescriptorContext? context, CultureInfo? culture, IEnumerable e)
-    {
-        PyObject pyList = new(CPythonAPI.PyList_New(0));
-
-        foreach (var item in e)
-        {
-            PyObject converted = ToPython(item, context, culture);
-            int result = CPythonAPI.PyList_Append(pyList.GetHandle(), converted!.GetHandle());
-            if (result == -1)
-            {
-                throw new Exception("Failed to set item in list");
-            }
-        }
-
-        return pyList;
-    }
-
-    private PyObject ConvertFromTuple(ITypeDescriptorContext? context, CultureInfo? culture, ITuple t)
-    {
-        List<PyObject> pyObjects = [];
-
-        for (var i = 0; i < t.Length; i++)
-        {
-            var currentValue = t[i];
-            if (currentValue is null)
-            {
-                // TODO: handle null values
-            }
-            pyObjects.Add(ToPython(currentValue, context, culture));
-        }
-
-        return PyTuple.CreateTuple(pyObjects);
-    }
 
     public override bool CanConvertTo(ITypeDescriptorContext? context, [NotNullWhen(true)] Type? destinationType) =>
         destinationType is not null && (
